+++ metatensor-torch/include/metatensor/torch/tensor.hpp
@@ -146,7 +146,8 @@ public:
     /// Move this `TensorMap` to the given `dtype` and `device`.
     TensorMap to(
         torch::optional<torch::Dtype> dtype = torch::nullopt,
-        torch::optional<torch::Device> device = torch::nullopt
+        torch::optional<torch::Device> device = torch::nullopt,
+        bool non_blocking = false
     ) const;
 
     /// Wrapper of the `to` function to enable using it with positional
@@ -162,7 +163,8 @@ public:
         torch::IValue positional_2,
         torch::optional<torch::Dtype> dtype,
         torch::optional<torch::Device> device,
-        torch::optional<std::string> arrays
+        torch::optional<std::string> arrays,
+        bool non_blocking = false
     ) const;
 
     /// Get the underlying metatensor TensorMap